requirements.txt
anthropic
python-docx
streamlit
python-dotenv